kafka_monitor/requirements.txt
kafka-python==2.0.2
requests
pymongo[srv]==3.11.2
sentry-sdk==1.0.0

# https://github.com/scrapy/scrapyd-client/pull/68
git+https://github.com/scrapy/scrapyd-client@d9a8871ff6ff09e7be24eba83ddec3e11c53b04f
wait-for-it==2.2.0
